11.07.2022

+++ cart/cart.py
@@ -43,7 +43,7 @@ class Cart:
         for product in products:
             cart[str(product.id)]['product'] = product
         for item in cart.values():
-            item['price'] = Decimal(item['price'])
+            item['price'] = float(item['price'])
             item['total_price'] = item['price'] * item['quantity']
             yield item
 
@@ -53,7 +53,7 @@ class Cart:
 
     def get_total_price(self):
         return sum(
-            Decimal(item['price']) * item['quantity']
+            float(item['price']) * item['quantity']
             for item in self.cart.values()
         )
